+++ src/api/product/apis.ts
@@ -2,7 +2,7 @@
 
 import { isNull, isUndefined } from "@/utils/type-guard";
 import { fetchWithToken } from "../fetchApi";
-import { createQueryString } from "@/utils/apiUtils";
+import { createQueryString } from "@/utils/apis/create-query-string";
 
 // 쿼리 파라미터 생성
 // const queryParams = new URLSearchParams({
@@ -30,7 +30,7 @@ export const getProductList = async (queryParams?: Record<string, string>) => {
 export const getProductDatail = async (id: string) => {
   try {
     const data = await fetchWithToken<GetProductDatailResponse>(
-      `product/${id}/`,
+      `product/${id}/`
     );
     if (isUndefined(data) || isNull(data)) {
       console.log("상품 데이터가 비어있습니다.");
@@ -48,7 +48,7 @@ export const getProductPriceGraph = async () => {
   const month = "3";
   try {
     const data = await fetchWithToken<GetProductDatailResponse>(
-      `price_history/?period=${month}&product_id=${id}/`,
+      `price_history/?period=${month}&product_id=${id}/`
     );
     if (isUndefined(data) || isNull(data)) {
       console.log("상품 데이터가 비어있습니다.");
@@ -64,7 +64,7 @@ export const getProductPriceGraph = async () => {
 export const getProductPriceInfo = async (id: string) => {
   try {
     const data = await fetchWithToken<GetProductInfoResponse>(
-      `product/${id}/price_info/`,
+      `product/${id}/price_info/`
     );
     if (isUndefined(data) || isNull(data)) {
       console.log("데이터가 비어있습니다.");
@@ -80,7 +80,7 @@ export const getProductPriceInfo = async (id: string) => {
 export const getRelatedProductList = async (id: string) => {
   try {
     const data = await fetchWithToken<GetRelatedProductListResponse[]>(
-      `product/${id}/related_product_list/`,
+      `product/${id}/related_product_list/`
     );
     if (isUndefined(data) || isNull(data)) {
       console.log("데이터가 비어있습니다.");
@@ -107,13 +107,10 @@ export const getProductListBySearch = async (
       },
       stringRecord
     );
-    if (isUndefined(data) || isNull(data) || isUndefined(data.results)) {
-      console.log("상품 데이터가 비어있습니다.");
-      return { results: [] };
-    }
+
     return data;
   } catch (error) {
     console.error("에러:", error);
-    return { results: [] };
+    return { count: 0, cursor: "", results: [] };
   }
 };
